Add GUI and update Makefile and code accordingly
A new struct "thread_args_t" has been typedefed in the regulator header. It is used as the argument to the functions called by the threads for the GUI and regulator. It is a much safer way to pass the arguments to the new threads. A variable "run" is also declared in the struct. It is used as the condition in the regulator loop and is set to 0 when the GUI exits.

The option to compile without the gui was added in the new Makefile. Make with target "nogui" to do so. The CFLAGS variable in the Makefile defines the macro NO_GUI which is checked in the run_gui header. If it is defined, a very small implementation of the function is made, without a loop. It sets the (new) regulator parameter "on" to 1, which enables a "continue" statement in the beginning of the regulator loop. The effect is that the loop runs continuously with no action. Be aware that setting this variable to 0 will not stop the regulator loop, simply stop its actions.

The thread demo was removed.

The gitignore was updated not to contain the executable.