Add nohw and nohworgui compile targets to Makefile
The nohw target will define the macro NO_HARDWARE which will make empty static function definitions of Kalman and all functions that are called in regulator.c so that the gui can be tested without those dependencies. The project can also be compiled without needing any of the dependencies using the nohworgui target. The static function defenitions are needed because both main.c and regulator.c includes regulator.h.

Compile support for qcustomplot is also added.