Implement real time sleep using clock_nanosleep
clock_nanosleep from time.h can be used for absolute sleep, which (contrary to a relative sleep) will not get an accumulative delay depending on preemptions by other threads.

Also moved dummy implementations for the nohw target from the regulator header to the source file since it is only regulator source file that needs the dummies.